style(test): use shared mocks
