+++ test/planner/widgets/planner_activities_header_test.dart
@@ -11,7 +11,6 @@ import 'package:go_router/go_router.dart';
 import 'package:mocktail/mocktail.dart';
 
 import '../../helpers/helpers.dart';
-import '../planner_mocks.dart';
 
 void main() {
   group('PlannerActivitiesHeader', () {
