+++ test/planner/widgets/planner_calendar_test.dart
@@ -6,7 +6,6 @@ import 'package:mocktail/mocktail.dart';
 import 'package:table_calendar/table_calendar.dart';
 
 import '../../helpers/helpers.dart';
-import '../planner_mocks.dart';
 
 void main() {
   group('PlannerCalendar', () {
