+++ test/planner/widgets/planner_fab_test.dart
@@ -10,7 +10,6 @@ import 'package:go_router/go_router.dart';
 import 'package:mockingjay/mockingjay.dart';
 
 import '../../helpers/helpers.dart';
-import '../planner_mocks.dart';
 
 void main() {
   group('PlannerFAB', () {
