+++ test/planner/widgets/planner_tabs_test.dart
@@ -8,7 +8,6 @@ import 'package:mocktail/mocktail.dart';
 import 'package:tasks_repository/tasks_repository.dart';
 
 import '../../helpers/helpers.dart';
-import '../planner_mocks.dart';
 
 void main() {
   group('PlannerTabs', () {
