+++ test/planner/widgets/planner_tasks_test.dart
@@ -7,7 +7,6 @@ import 'package:mocktail/mocktail.dart';
 import 'package:tasks_repository/tasks_repository.dart';
 
 import '../../helpers/helpers.dart';
-import '../planner_mocks.dart';
 
 void main() {
   group('PlannerTasks', () {
