+++ test/routine/routine_mocks.dart
@@ -1,9 +1,5 @@
-import 'package:bloc_test/bloc_test.dart';
 import 'package:flutter_planner/routine/routine.dart';
 import 'package:routines_api/routines_api.dart';
-
-class MockRoutineBloc extends MockBloc<RoutineEvent, RoutineState>
-    implements RoutineBloc {}
 
 final mockRoutine = Routine(
   userID: 'userID',
