+++ test/schedule/widgets/schedule_side_pane_test.dart
@@ -7,7 +7,6 @@ import 'package:mocktail/mocktail.dart';
 import 'package:routines_repository/routines_repository.dart';
 
 import '../../helpers/helpers.dart';
-import '../schedule_mocks.dart';
 
 void main() {
   group('ScheduleSidePane', () {
