--- test/sign_in/sign_in_mocks.dart
@@ -1,5 +1,0 @@
-import 'package:bloc_test/bloc_test.dart';
-import 'package:flutter_planner/sign_in/sign_in.dart';
-
-class MockSignInBloc extends MockBloc<SignInEvent, SignInState>
-    implements SignInBloc {}
