+++ test/sign_in/widgets/sign_in_error_test.dart
@@ -5,7 +5,6 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:mocktail/mocktail.dart';
 
 import '../../helpers/helpers.dart';
-import '../sign_in_mocks.dart';
 
 void main() {
   group('SignInError', () {
