--- test/sign_up/sign_up_mocks.dart
@@ -1,5 +1,0 @@
-import 'package:bloc_test/bloc_test.dart';
-import 'package:flutter_planner/sign_up/sign_up.dart';
-
-class MockSignUpBloc extends MockBloc<SignUpEvent, SignUpState>
-    implements SignUpBloc {}
