+++ test/task/task_mocks.dart
@@ -1,8 +1,4 @@
-import 'package:bloc_test/bloc_test.dart';
-import 'package:flutter_planner/task/task.dart';
 import 'package:tasks_repository/tasks_repository.dart';
-
-class MockTaskBloc extends MockBloc<TaskEvent, TaskState> implements TaskBloc {}
 
 final mockTask = Task(
   userID: 'userID',
