+++ test/task/widgets/task_text_field_test.dart
@@ -6,7 +6,6 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:mocktail/mocktail.dart';
 
 import '../../helpers/helpers.dart';
-import '../../planner/planner_mocks.dart';
 import '../task_mocks.dart';
 
 void main() {
